+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 * Fix SDRAM timing constraints
 * Fix SDRAM startup sequence
 * Add SDRAM write recovery delay
+* Register SDRAM output
+* Allow SDRAM refresh as soon as a request has completed
+* Fix timing issues in ROM controller
 
 ## v1.0.0-beta.0 (2019-09-22)